add bigdl model
